// frontend/src/http/post-estimate-ride/estimate-ride-type.ts
export type EstimateRideType = {
  routeResponse: any
  origin: {
    latitude: number
    longitude: number
  }
  destination: {
    latitude: number
    longitude: number
  }
  distance: number
  duration: string
  options: {
    value: number
    id: number
    name: string
    description: string
    vehicle: string
    review: {
      rating: number
      comment: string
    }
  }[]
}